pyproject.toml
[tool.poetry]
name = "pizzaria"
version = "0.1.0"
description = ""
authors = ["Patrick Rodrigues <pythrick@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
Django = "^3.2"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
uvicorn = "^0.13.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
